feat(#314): add javadocs
